gerenciamento_de_tarefas/gerenciamento_de_tarefas/src/requirements.txt
cx-Oracle==8.3.0
pandas==1.4.4
numpy==1.23.2
python-dateutil